refactor(controllers): add controllers into http layer
