+++ src/infra/http/controllers/treatment-type.controller.ts
