--- src/modules/user/controllers/user.controller.e2e-spec.ts
@@ -1,90 +1,0 @@
-import { faker } from '@faker-js/faker/locale/pt_BR';
-import type { INestApplication } from '@nestjs/common';
-import { Test } from '@nestjs/testing';
-import request from 'supertest';
-import { AppModule } from '@/app.module';
-import { PrismaService } from '@/modules/prisma/prisma.service';
-import {
-	makeAuthenticate,
-	makeUser,
-	makeUserProfessional,
-} from '@/test/factories';
-
-describe('User (E2E)', () => {
-	let app: INestApplication;
-	let prisma: PrismaService;
-
-	beforeAll(async () => {
-		const moduleRef = await Test.createTestingModule({
-			imports: [AppModule],
-		}).compile();
-
-		app = moduleRef.createNestApplication();
-		prisma = moduleRef.get<PrismaService>(PrismaService);
-
-		await app.init();
-	});
-
-	it('[GET] /accounts/me', async () => {
-		const user = await makeUser(prisma);
-		const token = await makeAuthenticate(app, user.email);
-
-		const getUser = await request(app.getHttpServer())
-			.get('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		expect(getUser.statusCode).toBe(200);
-		expect(getUser.body).toHaveProperty('name');
-	});
-	it('[POST] /auth/signin - should return 401 when user try to login with wrong role', async () => {
-		const user = await makeUserProfessional(prisma);
-
-		await request(app.getHttpServer()).post('/auth/signup');
-
-		const token = await makeAuthenticate(app, user.email);
-
-		const response = await request(app.getHttpServer())
-			.get('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		expect(response.status).toBe(401);
-	});
-	it('[DELETE] /accounts/me', async () => {
-		const user = await makeUser(prisma);
-		const token = await makeAuthenticate(app, user.email);
-
-		const userExists = await request(app.getHttpServer())
-			.get('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		const deleteUser = await request(app.getHttpServer())
-			.delete('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		const isUserDeleted = await request(app.getHttpServer())
-			.get('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		expect(userExists.statusCode).toBe(200);
-		expect(deleteUser.statusCode).toBe(204);
-		expect(isUserDeleted.statusCode).toBe(404);
-	});
-
-	it('[PATCH] /accounts/me', async () => {
-		const user = await makeUser(prisma);
-		const token = await makeAuthenticate(app, user.email);
-		const uniqueEmail = faker.internet.email(); // email único para cada vez que rodar o teste
-
-		const userExists = await request(app.getHttpServer())
-			.get('/accounts/me')
-			.set('Authorization', `Bearer ${token}`);
-
-		const updateUserProfile = await request(app.getHttpServer())
-			.patch('/accounts/me')
-			.set('Authorization', `Bearer ${token}`)
-			.send({ name: 'Novo Nome', email: uniqueEmail }); //passa os campos que serão atualizados
-
-		expect(userExists.statusCode).toBe(200);
-		expect(updateUserProfile.statusCode).toBe(200);
-	});
-});
